demo/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-extensions==2.2.5
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
pytz==2019.3
requests==2.22.0
six==1.13.0
-e git+https://github.com/techoutlooks/django-smart-models.git@2c6cf6a0e0f64c067db30853d6debb3cf4111bfe#egg=smartmodels
sqlparse==0.3.0
urllib3==1.25.7
